+++ packages/banner/src/Banner.tsx
@@ -1,296 +1,26 @@
 import React from 'react';
-import ImportantWithCircleIcon from '@leafygreen-ui/icon/dist/ImportantWithCircle';
-import InfoWithCircleIcon from '@leafygreen-ui/icon/dist/InfoWithCircle';
-import WarningIcon from '@leafygreen-ui/icon/dist/Warning';
-import CheckmarkWithCircleIcon from '@leafygreen-ui/icon/dist/CheckmarkWithCircle';
+import PropTypes from 'prop-types';
 import XIcon from '@leafygreen-ui/icon/dist/X';
-import { css, cx } from '@leafygreen-ui/emotion';
-import { palette } from '@leafygreen-ui/palette';
-import { HTMLElementProps } from '@leafygreen-ui/lib';
+import { cx } from '@leafygreen-ui/emotion';
 import IconButton from '@leafygreen-ui/icon-button';
-
-const Variant = {
-  Info: 'info',
-  Warning: 'warning',
-  Danger: 'danger',
-  Success: 'success',
-} as const;
-
-type Variant = typeof Variant[keyof typeof Variant];
-
-export { Variant };
-
-const defaultBorderSpacing = 12;
-
-const baseBannerStyles = css`
-  position: relative;
-  display: flex;
-  min-height: 40px;
-  padding: 10px 12px 10px 20px;
-  border-width: 1px 1px 1px 0px;
-  border-style: solid;
-  border-radius: 12px;
-  font-size: 13px;
-  line-height: 20px;
-  font-family: 'Euclid Circular A', Helvetica, Arial, sans-serif; // TODO: Refresh – remove when fonts are updated
-
-  &:before {
-    content: '';
-    position: absolute;
-    width: 13px;
-    top: -1px;
-    bottom: -1px;
-    left: 0px;
-    border-radius: 12px 0px 0px 12px;
-  }
-`;
-
-//TODO: refresh - remove when IconButton focus styles are updated
-const baseDismissButtonStyles = css`
-  &:focus {
-    color: ${palette.gray.dark3};
-    outline: 2px solid ${palette.blue.light1};
-    border: 2px solid ${palette.white};
-  }
-`;
-
-const flexShrink = css`
-  flex-shrink: 0;
-`;
-
-const cursorPointer = css`
-  cursor: pointer;
-`;
-
-const renderedImageStyles = css`
-  // this margin is set to control text alignment with the base of the renderedImage
-  margin-top: 3px;
-  margin-bottom: 3px;
-
-  width: 32px;
-  height: 32px;
-  flex-shrink: 0;
-`;
-
-type StyledElements = 'body' | 'dismissButton';
-
-const bannerVariantStyles: Record<Variant, Record<StyledElements, string>> = {
-  [Variant.Info]: {
-    body: css`
-      color: ${palette.blue.dark2};
-      border-color: ${palette.blue.light2};
-      border-left-color: ${palette.blue.base};
-      background-color: ${palette.blue.light3};
-
-      &:before {
-        background: linear-gradient(
-          to left,
-          transparent 6px,
-          ${palette.blue.base} 6px
-        );
-      }
-    `,
-    dismissButton: css`
-      color: ${palette.blue.dark2};
-
-      &:active,
-      &:hover {
-        color: ${palette.blue.dark2};
-
-        &:before {
-          background-color: ${palette.blue.light2};
-        }
-      }
-
-      &:focus {
-        &:before {
-          background-color: ${palette.blue.light2};
-        }
-      }
-    `,
-  },
-
-  [Variant.Warning]: {
-    body: css`
-      color: ${palette.yellow.dark2};
-      border-color: ${palette.yellow.light2};
-      border-left-color: ${palette.yellow.base};
-      background-color: ${palette.yellow.light3};
-
-      &:before {
-        background: linear-gradient(
-          to left,
-          transparent 6px,
-          ${palette.yellow.base} 6px
-        );
-      }
-    `,
-    dismissButton: css`
-      color: ${palette.yellow.dark2};
-
-      &:active,
-      &:hover {
-        color: ${palette.yellow.dark2};
-
-        &:before {
-          background-color: ${palette.yellow.light2};
-        }
-      }
-
-      &:focus {
-        &:before {
-          background-color: ${palette.yellow.light2};
-        }
-      }
-    `,
-  },
-
-  [Variant.Danger]: {
-    body: css`
-      color: ${palette.red.dark2};
-      border-color: ${palette.red.light2};
-      border-left-color: ${palette.red.base};
-      background-color: ${palette.red.light3};
-
-      &:before {
-        background: linear-gradient(
-          to left,
-          transparent 6px,
-          ${palette.red.base} 6px
-        );
-      }
-    `,
-    dismissButton: css`
-      color: ${palette.red.dark2};
-
-      &:active,
-      &:hover {
-        color: ${palette.red.dark2};
-
-        &:before {
-          background-color: ${palette.red.light2};
-        }
-      }
-
-      &:focus {
-        &:before {
-          background-color: ${palette.red.light2};
-        }
-      }
-    `,
-  },
-
-  [Variant.Success]: {
-    body: css`
-      color: ${palette.green.dark2};
-      border-color: ${palette.green.light2};
-      border-left-color: ${palette.green.base};
-      background-color: ${palette.green.light3};
-
-      &:before {
-        background: linear-gradient(
-          to left,
-          transparent 6px,
-          ${palette.green.dark1} 6px
-        );
-      }
-    `,
-    dismissButton: css`
-      color: ${palette.green.dark2};
-
-      &:active,
-      &:hover {
-        color: ${palette.green.dark2};
-
-        &:before {
-          background-color: ${palette.green.light2};
-        }
-      }
-
-      &:focus {
-        &:before {
-          background-color: ${palette.green.light2};
-        }
-      }
-    `,
-  },
-} as const;
-
-const map = {
-  [Variant.Info]: { color: palette.blue.base, icon: InfoWithCircleIcon },
-  [Variant.Warning]: {
-    color: palette.yellow.dark2,
-    icon: ImportantWithCircleIcon,
-  },
-  [Variant.Danger]: { color: palette.red.base, icon: WarningIcon },
-  [Variant.Success]: {
-    color: palette.green.dark1,
-    icon: CheckmarkWithCircleIcon,
-  },
-};
-
-const getTextStyle = (image: boolean, dismissible: boolean) => {
-  const defaultIconSize = 16;
-
-  const styleObj: {
-    marginLeft?: string;
-    marginRight?: string;
-  } = {
-    marginLeft: undefined,
-    marginRight: undefined,
-  };
-
-  if (image) {
-    styleObj.marginLeft = '17px';
-    styleObj.marginRight = '4px';
-    if (dismissible) {
-      styleObj.marginRight = `${defaultIconSize + defaultBorderSpacing}px`;
-    }
-  } else {
-    styleObj.marginLeft = `13px`;
-    styleObj.marginRight = '10px';
-    if (dismissible) {
-      styleObj.marginRight = `${defaultIconSize + 16}px`;
-    }
-  }
-
-  return css`
-    align-self: center;
-    flex-grow: 1;
-    margin-left: ${styleObj.marginLeft};
-    margin-right: ${styleObj.marginRight};
-  `;
-};
-
-interface BannerProps extends HTMLElementProps<'div', never> {
-  /**
-   * Sets the variant for the Banner
-   *
-   * Default: `'info'`
-   */
-  variant?: Variant;
-
-  /**
-   * Illustration that will replace default Icon when the prop is supplied
-   *
-   * @type ReactElement
-   */
-  image?: React.ReactElement;
-
-  /**
-   * Determines whether or not the Banner is dismissible
-   *
-   * Default: `false`
-   */
-  dismissible?: boolean;
-
-  /**
-   * Callback fired when dismiss button is clicked
-   *
-   * Default: `() => {}`
-   */
-  onClose?: React.MouseEventHandler;
-}
+import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
+import { BaseFontSize } from '@leafygreen-ui/tokens';
+import {
+  useUpdatedBaseFontSize,
+  bodyTypeScaleStyles,
+} from '@leafygreen-ui/typography';
+import { BannerProps, Variant } from './types';
+import {
+  bannerIconPositionStyles,
+  bannerVariantStyles,
+  baseBannerStyles,
+  dismissibleIconStyles,
+  getTextStyle,
+  iconStyles,
+  map,
+  renderedImageStyles,
+  bannerDismissibleStyles,
+} from './styles';
 
 /**
  *
@@ -300,6 +30,8 @@ interface BannerProps extends HTMLElementProps<'div', never> {
  * @param props.image Illustration that will replace default Icon when the prop is supplied.
  * @param props.dismissible Determines whether or not the Banner is dismissible.
  * @param props.onClose Callback fired when dismiss button is clicked.
+ * @param props.darkMode Determines whether or not the component will be rendered in dark mode.
+ * @param props.baseFontSize The base font size of the title and text rendered in children.
  */
 export default function Banner({
   variant = Variant.Info,
@@ -308,9 +40,13 @@ export default function Banner({
   image,
   children,
   className,
+  darkMode: darkModeProp,
+  baseFontSize: baseFontSizeProp,
   ...rest
 }: BannerProps) {
-  const { icon: Icon, color } = map[variant];
+  const { theme, darkMode } = useDarkMode(darkModeProp);
+  const baseFontSize = useUpdatedBaseFontSize(baseFontSizeProp);
+  const { icon: Icon, color } = map[theme][variant];
 
   const renderIcon = image ? (
     React.cloneElement(image, {
@@ -319,12 +55,7 @@ export default function Banner({
   ) : (
     <Icon
       fill={color}
-      className={cx(
-        flexShrink,
-        css`
-          margin-top: 3px;
-        `,
-      )}
+      className={cx(iconStyles, bannerIconPositionStyles[baseFontSize])}
     />
   );
 
@@ -333,7 +64,11 @@ export default function Banner({
       role="alert"
       className={cx(
         baseBannerStyles,
-        bannerVariantStyles[variant].body,
+        bodyTypeScaleStyles[baseFontSize],
+        bannerVariantStyles[theme][variant].body,
+        {
+          [bannerDismissibleStyles]: dismissible,
+        },
         className,
       )}
       {...rest}
@@ -343,17 +78,12 @@ export default function Banner({
       {dismissible && (
         <IconButton
           className={cx(
-            flexShrink,
-            cursorPointer,
-            css`
-              margin-top: -3px;
-              left: 8px;
-            `,
-            bannerVariantStyles[variant].dismissButton,
-            baseDismissButtonStyles,
+            dismissibleIconStyles,
+            bannerVariantStyles[theme][variant].dismissButton,
           )}
           aria-label="Close Message"
           onClick={onClose}
+          darkMode={darkMode}
         >
           <XIcon />
         </IconButton>
@@ -361,3 +91,14 @@ export default function Banner({
     </div>
   );
 }
+
+Banner.displayName = 'Banner';
+
+Banner.propTypes = {
+  darkMode: PropTypes.bool,
+  variant: PropTypes.oneOf(Object.values(Variant)),
+  onClose: PropTypes.func,
+  dismissable: PropTypes.bool,
+  image: PropTypes.element,
+  baseFontSize: PropTypes.oneOf(Object.values(BaseFontSize)),
+};