+++ packages/table/src/styles.tsx
@@ -5,7 +5,7 @@ const sharedStyles = css`
   min-width: 40px;
   padding: var(--lg-cell-padding-block) 8px;
   box-sizing: border-box;
-  vertical-align: middle;
+  vertical-align: baseline;
   text-align: left; // Justification is updated in \`Row.tsx\` for number cells
 `;