+++ packages/palette/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leafygreen-ui/palette
 
+## 3.4.2
+
+### Patch Changes
+
+- 8d7534e9: Adds `tsdoc.json` to published package files
+
 ## 3.4.1
 
 ### Patch Changes