+++ packages/side-nav/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @leafygreen-ui/side-nav
 
+## 10.0.2
+
+### Patch Changes
+
+- 8d7534e9: Adds `tsdoc.json` to published package files
+- Updated dependencies [a1202635]
+- Updated dependencies [7caa1c3e]
+- Updated dependencies [e39d8469]
+- Updated dependencies [8d7534e9]
+- Updated dependencies [5de9d6ad]
+- Updated dependencies [1e708bd3]
+  - @leafygreen-ui/menu@16.0.0
+  - @leafygreen-ui/icon@11.10.2
+  - @leafygreen-ui/typography@13.1.2
+  - @leafygreen-ui/a11y@1.3.1
+  - @leafygreen-ui/box@3.0.8
+  - @leafygreen-ui/emotion@4.0.1
+  - @leafygreen-ui/hooks@7.3.1
+  - @leafygreen-ui/leafygreen-provider@2.3.3
+  - @leafygreen-ui/lib@9.4.2
+  - @leafygreen-ui/palette@3.4.2
+  - @leafygreen-ui/portal@4.0.4
+  - @leafygreen-ui/tokens@1.3.2
+  - @leafygreen-ui/tooltip@7.1.1
+
 ## 10.0.1
 
 ### Patch Changes